Resolve PR feedback
From the source:

```
There are a couple corner cases that I'd like to make sure we don't miss here. One is when the range is descending, e.g. `(5..0)`. In that case I think we should raise an error, since it's unclear what the intended behavior is.

I'm also a little worried about ranges that exclude the end value, e.g. `(0...5)`, since most databases use inclusive ranges for `BETWEEN`. In the case of a `Range` that excludes the end value (see [#exclude_end?](https://ruby-doc.org/core-2.2.0/Range.html#method-i-exclude_end-3F)), can you change this to use an explicit `>=` and `<`? I think a test with a range of `(0...5)` that ensures that `4.5` is included but `5` is excluded would be great.
```