Use a Arel::Nodes::Case node
Using Arel a little more heavily gets free escaping, which means you don't need to worry so much about SQL injection or things like that. Also it means less manual creation of SQL, which is always better. This also opens the door for a little more validation in the future.