Fix exclusive range arel usage
With newer versions of arel, the `in` method enumerates the range and
checks if the value is in that enumeration. The behavior we want is to
find out if the values are within the range, regardless of whether or
not they are enumerated. We use `#between` to get this behavior.